# libs/api-contract/bikes-api/swagger.yaml
basePath: /
definitions:
  domain.Bike:
    properties:
      created_at:
        type: string
      current_owner_id:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/domain.BikeImage'
        type: array
      location_city:
        type: string
      make_model:
        type: string
      price:
        type: number
      serial_number:
        type: string
      status:
        $ref: '#/definitions/domain.BikeStatus'
      updated_at:
        type: string
      year:
        type: integer
    type: object
  domain.BikeImage:
    properties:
      bike_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  domain.BikeStatus:
    enum:
    - registered
    - for_sale
    - stolen
    - transferred
    type: string
    x-enum-varnames:
    - StatusRegistered
    - StatusForSale
    - StatusStolen
    - StatusTransferred
  handler.ConfirmUploadRequest:
    properties:
      object_key:
        type: string
    required:
    - object_key
    type: object
  handler.RegisterBikeRequest:
    properties:
      description:
        type: string
      location_city:
        type: string
      make_model:
        type: string
      price:
        type: number
      serial_number:
        type: string
      year:
        type: integer
    type: object
  handler.UploadURLRequest:
    properties:
      filename:
        type: string
    required:
    - filename
    type: object
  handler.UploadURLResponse:
    properties:
      object_key:
        type: string
      upload_url:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: High-trust Bicycle Registry & Marketplace API.
  title: VeloTrace Bikes API
  version: "1.0"
paths:
  /admin/bikes:
    get:
      description: Returns every bike in the system. Strictly for users with admin
        role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Bike'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List all bikes (Admin)
      tags:
      - admin
  /bikes:
    get:
      description: Returns a list of bikes with status 'for_sale'. Sensitive fields
        are redacted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Bike'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List public bikes for sale
      tags:
      - bikes
    post:
      consumes:
      - application/json
      description: Creates a bike entry and an ownership record in a single transaction
      parameters:
      - description: Bike registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterBikeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Bike'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Register a new bike
      tags:
      - bikes
  /bikes/{id}:
    get:
      description: 'Returns bike details. Hybrid logic: Public fields if for_sale,
        full fields if owner/admin, 404 otherwise.'
      parameters:
      - description: Bike ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Bike'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get bike by ID
      tags:
      - bikes
  /bikes/{id}/images/confirm:
    post:
      consumes:
      - application/json
      description: Creates a bike_image record. Sets as primary if it's the first
        image.
      parameters:
      - description: Bike ID
        in: path
        name: id
        required: true
        type: string
      - description: Confirm upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ConfirmUploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Confirm image upload
      tags:
      - images
  /bikes/{id}/upload-url:
    post:
      consumes:
      - application/json
      description: Generates a unique object key and a presigned URL valid for 15
        minutes
      parameters:
      - description: Bike ID
        in: path
        name: id
        required: true
        type: string
      - description: Upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UploadURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UploadURLResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get presigned upload URL
      tags:
      - images
  /my/bikes:
    get:
      description: Returns all bikes owned by the authenticated user with full metadata.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Bike'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List my bikes
      tags:
      - bikes
swagger: "2.0"